commitSun09:20

+++ Infrastructure/Components/QtyBox.cs
@@ -13,13 +13,12 @@ namespace Infrastructure
         private IWebElement QtyValue => ParentElement.WaitAndFindElement(By.CssSelector(".cart_quantity_input.form-control.grey"));
         private IWebElement DownButton => ParentElement.WaitAndFindElement(By.CssSelector(".cart_quantity_down.btn.btn-default.button-minus"));
         private IWebElement UpButton => ParentElement.WaitAndFindElement(By.CssSelector(".cart_quantity_up.btn.btn-default.button-plus"));
-        //check these selectors
 
         public QtyBox(IWebDriver driver, IWebElement parentElement) : base(driver, parentElement)
         {
         }
 
-        public int GetQtyValue() => int.Parse(QtyValue.GetAttribute("value"));
+        public double GetQtyValue() => double.Parse(QtyValue.GetAttribute("value"));
 
         public CartPage ClickOnUpButton()
         {
@@ -27,12 +26,14 @@ namespace Infrastructure
 
             return new CartPage(Driver);
         }
+
         public CartPage ClickDownUpButton()
         {
             DownButton.Click();
 
             return new CartPage(Driver);
         }
+
         public CartPage ChangeQty(double number)
         {
             QtyValue.Clear();
